# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/experiments/baseinfo/update-info": {
            "put": {
                "description": "根据实验ID更新实验基本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实验"
                ],
                "summary": "更新实验基本信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "实验ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "实验基本信息",
                        "name": "info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.UpdateExperimentBaseInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.UpdateExperimentBaseInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/experiments/create-experiment": {
            "post": {
                "description": "创建一个新的实验",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实验"
                ],
                "summary": "创建新实验",
                "parameters": [
                    {
                        "description": "实验信息",
                        "name": "experiment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.CreateExperimentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.CreateExperimentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/experiments/delete-experiment": {
            "delete": {
                "description": "根据实验ID删除实验",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实验"
                ],
                "summary": "删除实验",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "实验ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.DeleteExperimentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/experiments/delete-gropup": {
            "delete": {
                "description": "删除指定实验中的特定实验组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实验"
                ],
                "summary": "删除实验组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "实验ID",
                        "name": "experiment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "实验组ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.DeleteExperimentGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/experiments/getinfo": {
            "get": {
                "description": "根据实验ID获取实验详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实验"
                ],
                "summary": "获取实验信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "实验ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.GetExperimentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/experiments/groups/create-groups": {
            "post": {
                "description": "向现有实验添加一个新的实验组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "实验"
                ],
                "summary": "添加实验组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "实验ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "实验组信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.AddExperimentGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb_experiment.AddExperimentGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb_experiment.AddExperimentGroupRequest": {
            "type": "object",
            "properties": {
                "experiment_id": {
                    "type": "integer"
                },
                "group": {
                    "$ref": "#/definitions/pb_experiment.ExperimentGroup"
                }
            }
        },
        "pb_experiment.AddExperimentGroupResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_experiment.CreateExperimentRequest": {
            "type": "object",
            "properties": {
                "experiment_info": {
                    "$ref": "#/definitions/pb_experiment.ExperimentInfo"
                }
            }
        },
        "pb_experiment.CreateExperimentResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "experiment_id": {
                    "type": "integer"
                },
                "experiment_info": {
                    "$ref": "#/definitions/pb_experiment.ExperimentInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_experiment.DeleteExperimentGroupResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_experiment.DeleteExperimentResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_experiment.ExperimentGroup": {
            "type": "object",
            "properties": {
                "allocation": {
                    "description": "实验组占实验比例",
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "from_experiment_id": {
                    "description": "实验组来源实验ID",
                    "type": "integer"
                },
                "group_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "whitelist_userpackageID": {
                    "description": "白名单",
                    "type": "integer"
                }
            }
        },
        "pb_experiment.ExperimentInfo": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "experiment_id": {
                    "type": "integer"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb_experiment.ExperimentGroup"
                    }
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/pb_experiment.ExperimentStatus"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "pb_experiment.ExperimentStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-comments": {
                "ExperimentStatus_EXPERIMENT_STATUS_CLOSED": "关闭状态",
                "ExperimentStatus_EXPERIMENT_STATUS_READYING": "准备状态",
                "ExperimentStatus_EXPERIMENT_STATUS_RUNNING": "灰度中",
                "ExperimentStatus_EXPERIMENT_STATUS_STOPPED": "停止状态",
                "ExperimentStatus_EXPERIMENT_STATUS_UNKNOWN": "未知状态",
                "ExperimentStatus_ExPERIMENT_STATUS_LAUNCHED": "launched状态"
            },
            "x-enum-varnames": [
                "ExperimentStatus_EXPERIMENT_STATUS_UNKNOWN",
                "ExperimentStatus_EXPERIMENT_STATUS_READYING",
                "ExperimentStatus_EXPERIMENT_STATUS_RUNNING",
                "ExperimentStatus_EXPERIMENT_STATUS_STOPPED",
                "ExperimentStatus_EXPERIMENT_STATUS_CLOSED",
                "ExperimentStatus_ExPERIMENT_STATUS_LAUNCHED"
            ]
        },
        "pb_experiment.GetExperimentResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "experiment_info": {
                    "$ref": "#/definitions/pb_experiment.ExperimentInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "pb_experiment.UpdateExperimentBaseInfoRequest": {
            "type": "object",
            "properties": {
                "experiment_info": {
                    "$ref": "#/definitions/pb_experiment.ExperimentInfo"
                }
            }
        },
        "pb_experiment.UpdateExperimentBaseInfoResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}